requirements/base.txt
# Base
pytz==2020.1
python-slugify>=4.0.0,<4.1.0
Pillow>=7.1.2,<7.2.0
psycopg2>=2.8.6,<2.9.0 # --no-binary
# psycopg2-binary

# Django
django>=3.0.7,<3.1.0

# Environment
django-environ==0.4.5

# Django REST Framework
djangorestframework>=3.11.0,<3.12.0
djangorestframework-simplejwt==5.0.0

# Django REST Reset Passwords
django-rest-passwordreset>=1.1.0,<1.2.0

# Django CORS
django-cors-headers>=3.4.0,<3.5.0

# JWT
pyjwt

# Passwords security
argon2-cffi>=20.1.0,<20.2.0

# Static files
whitenoise>=5.1.0,<5.2.0

# Celery
redis>=3.5.3,<3.6.0
django-redis>=4.12.1,<4.13.0
celery>=4.4.5,<4.5.0
flower>=0.9.4,<1.0.0
